requirements.txt
scikit-image
tensorflow==1.4.0
keras>=2.0.8
opencv-python
matplotlib
numpy
gitpython
shapely